modify support profile command
