+++ cmd/support-profile-stop.go
@@ -18,13 +18,7 @@
 package cmd
 
 import (
-	"io"
-	"io/ioutil"
-	"os"
-	"time"
-
 	"github.com/minio/cli"
-	"github.com/minio/mc/pkg/probe"
 	"github.com/minio/pkg/console"
 )
 
@@ -36,100 +30,11 @@ var supportProfileStopCmd = cli.Command{
 	Before:          setGlobalsFromContext,
 	Flags:           globalFlags,
 	HideHelpCommand: true,
-	CustomHelpTemplate: `NAME:
-  {{.HelpName}} - {{.Usage}}
-
-USAGE:
-  {{.HelpName}} [FLAGS] TARGET
-
-FLAGS:
-  {{range .VisibleFlags}}{{.}}
-  {{end}}
-EXAMPLES:
-    2. Download latest profile data in the current directory
-       {{.Prompt}} {{.HelpName}} myminio/
-`,
-}
-
-func checkAdminProfileStopSyntax(ctx *cli.Context) {
-	if len(ctx.Args()) != 1 {
-		cli.ShowCommandHelpAndExit(ctx, "stop", 1) // last argument is exit code
-	}
-}
-
-// moveFile - os.Rename cannot handle cross device renames, in our situation
-// it is possible that /tmp is mounted from a separate partition and current
-// working directory is a different partition. To allow all situations to
-// be handled appropriately use this function instead of os.Rename()
-func moveFile(sourcePath, destPath string) error {
-	inputFile, err := os.Open(sourcePath)
-	if err != nil {
-		return err
-	}
-
-	outputFile, err := os.Create(destPath)
-	if err != nil {
-		inputFile.Close()
-		return err
-	}
-	defer outputFile.Close()
-
-	_, err = io.Copy(outputFile, inputFile)
-	inputFile.Close()
-	if err != nil {
-		return err
-	}
-
-	// The copy was successful, so now delete the original file
-	return os.Remove(sourcePath)
+	Hidden:          true,
 }
 
 // mainSupportProfileStop - the entry function of profile stop command
 func mainSupportProfileStop(ctx *cli.Context) error {
-	// Check for command syntax
-	checkAdminProfileStopSyntax(ctx)
-
-	// Get the alias parameter from cli
-	args := ctx.Args()
-	aliasedURL := args.Get(0)
-
-	// Create a new MinIO Admin Client
-	client, err := newAdminClient(aliasedURL)
-	if err != nil {
-		fatalIf(err.Trace(aliasedURL), "Unable to initialize admin client.")
-		return nil
-	}
-
-	// Create profile zip file
-	tmpFile, e := ioutil.TempFile("", "mc-profile-")
-	fatalIf(probe.NewError(e), "Unable to download profile data.")
-
-	// Ask for profile data, which will come compressed with zip format
-	zippedData, adminErr := client.DownloadProfilingData(globalContext)
-	fatalIf(probe.NewError(adminErr), "Unable to download profile data.")
-
-	// Copy zip content to target download file
-	_, e = io.Copy(tmpFile, zippedData)
-	fatalIf(probe.NewError(e), "Unable to download profile data.")
-
-	// Close everything
-	zippedData.Close()
-	tmpFile.Close()
-
-	downloadPath := "profile.zip"
-
-	fi, e := os.Stat(downloadPath)
-	if e == nil && !fi.IsDir() {
-		e = moveFile(downloadPath, downloadPath+"."+time.Now().Format(dateTimeFormatFilename))
-		fatalIf(probe.NewError(e), "Unable to create a backup of profile.zip")
-	} else {
-		if !os.IsNotExist(e) {
-			fatal(probe.NewError(e), "Unable to download profile data.")
-		}
-	}
-
-	fatalIf(probe.NewError(moveFile(tmpFile.Name(), downloadPath)), "Unable to download profile data.")
-
-	console.Infof("Profile data successfully downloaded as %s\n", downloadPath)
+	console.Infoln("Please use 'mc support profile <alias>'")
 	return nil
 }
